+++ radiospectra/net/sources/ilofar.py
@@ -3,8 +3,8 @@ import numpy as np
 import astropy.units as u
 from sunpy.net import attrs as a
 from sunpy.net.dataretriever.client import GenericClient, QueryResponse
+from sunpy.net.scraper import Scraper
 from sunpy.time import TimeRange
-from sunpy.util.scraper import Scraper
 
 from radiospectra.net.attrs import PolType