+++ radiospectra/net/sources/tests/test_wind_client.py
@@ -18,6 +18,7 @@ def client():
     return WAVESClient()
 
 
+@pytest.mark.skip("Upstream data server removed")
 @pytest.mark.remote_data
 def test_fido():
     atr = a.Time("2010/10/01", "2010/10/02")