+++ radiospectra/net/sources/wind.py
@@ -28,8 +28,8 @@ class WAVESClient(GenericClient):
     >>> import radiospectra.net
     >>> from sunpy.net import Fido, attrs as a
     >>> results = Fido.search(a.Time("2010/10/01", "2010/10/02"),
-    ...                       a.Instrument('WAVES'))  # doctest: +REMOTE_DATA
-    >>> results #doctest: +REMOTE_DATA
+    ...                       a.Instrument('WAVES'))  # doctest: +REMOTE_DATA +SKIP
+    >>> results #doctest: +REMOTE_DATA +SKIP
     <sunpy.net.fido_factory.UnifiedResponse object at ...>
     Results from 1 Provider:
     <BLANKLINE>